requirements.txt
flask
flask_sqlalchemy
flask_login
pandas
numpy
werkzeug.security
